+++ assets/_src/modules/full-offload.tsx
@@ -1,6 +1,7 @@
 /**
  * External dependencies
  */
+import { MouseEvent, useContext, useState } from 'react';
 import { mdiImageMultipleOutline } from '@mdi/js';
 
 /**
@@ -11,9 +12,20 @@ import { __ } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
+import SettingsContext from '../context/settings';
 import Card from '../components/card';
+import ProgressBar from '../components/progress';
 
 const FullOffload = () => {
+	const [action, setAction] = useState('');
+	const { inProgress, setInProgress } = useContext(SettingsContext);
+
+	const runAction = (e: MouseEvent, actionName: string) => {
+		e.preventDefault();
+		setAction(actionName);
+		setInProgress(true);
+	};
+
 	return (
 		<Card
 			icon={mdiImageMultipleOutline}
@@ -30,6 +42,33 @@ const FullOffload = () => {
 				<p>
 					{__(
 						'By enabling this feature, you understand the potential risks of removing media files from the media library. Please ensure you have a backup of your media library.',
+						'cf-images'
+					)}
+				</p>
+
+				{inProgress && <ProgressBar action={action} />}
+
+				<div className="buttons mb-1">
+					<button
+						className="button is-small"
+						onClick={(e) => runAction(e, 'full-remove')}
+						disabled={inProgress}
+					>
+						{__('Bulk delete', 'cf-images')}
+					</button>
+
+					<button
+						className="button is-small"
+						onClick={(e) => runAction(e, 'full-restore')}
+						disabled={inProgress}
+					>
+						{__('Bulk restore', 'cf-images')}
+					</button>
+				</div>
+
+				<p>
+					{__(
+						'* Bulk actions will remove/restore physical files from the WordPress media library. Images that have not been offloaded to Cloudflare will be skipped.',
 						'cf-images'
 					)}
 				</p>