+++ assets/_src/routes/cloudflare/settings.tsx
@@ -22,6 +22,7 @@ import CloudflareDisconnect from '../../modules/actions/cf-disconnect';
 import CloudflareStats from '../../modules/cloudflare/cf-stats';
 import Logging from '../../modules/cloudflare/logging';
 import Service from '../../modules/cloudflare/service';
+import RSSFeeds from '../../modules/cloudflare/rss-feeds';
 
 /**
  * Cloudflare Images settings routes.
@@ -52,6 +53,7 @@ const CloudflareSettings = () => {
 				<CustomDomain />
 				<PageParser />
 				<DisableAsync />
+				<RSSFeeds />
 				<Service />
 				<Logging />
 				<CloudflareDisconnect />