+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+= 1.8.0 - 16.02.2024 =
+
+Added:
+* Support for RSS feeds
+* Auto crop option. If an image width matches the image height - auto crop the image.
+* Use img width size. New option that allows using the img width attribute value for the image size, if the value is smaller than the requested image.
+* Bulk remove files from the media library.
+* Bulk restore files to the media library.
+
+Fixed:
+* Page parser will now add the wp-image-* class to images that do not have it
+* Page parser not detecting images that have a custom title set in the media library
+
 = 1.7.1 - 31.12.2023 =
 
 Fixed: